--- app/src/main/java/com/whitebeach/atleticolineupapp/app/component/dragDrop/DragData.kt
@@ -1,8 +1,0 @@
-package com.whitebeach.atleticolineupapp.app.component.dragDrop
-
-import com.whitebeach.atleticolineupapp.dragAndDrop.MimeType
-
-class DragData (
-    val type: MimeType = MimeType.TEXT_PLAIN,
-    val data: Any? = null
-)
