--- app/src/main/java/com/whitebeach/atleticolineupapp/app/theme/Color.kt
@@ -1,10 +1,0 @@
-package com.whitebeach.atleticolineupapp.app.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple200 = Color(0xFFBB86FC)
-val Purple500 = Color(0xFF6200EE)
-val Purple700 = Color(0xFF3700B3)
-val Teal200 = Color(0xFF03DAC5)
-val MidNightBlue = Color(0xFF191970)
-val Red = Color(0xFFF3002A)
