--- app/src/main/java/com/whitebeach/atleticolineupapp/app/theme/Theme.kt
@@ -1,47 +1,0 @@
-package com.whitebeach.atleticolineupapp.app.theme
-
-import androidx.compose.foundation.isSystemInDarkTheme
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Typography
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.material3.lightColorScheme
-import androidx.compose.runtime.Composable
-
-private val darkColorPalette = darkColorScheme(
-    primary = Red,
-    secondary  = Purple700,
-    tertiary  = Teal200
-)
-
-private val lightColorPalette = lightColorScheme(
-    primary = Red,
-    secondary  = Red,
-    tertiary  = Red
-
-    /* Other default colors to override
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black,
-    */
-)
-
-@Composable
-fun AtleticoLineupAppTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    content: @Composable () -> Unit
-) {
-    val colors = if (darkTheme) {
-        darkColorPalette
-    } else {
-        lightColorPalette
-    }
-
-    MaterialTheme(
-        colorScheme = colors,
-        typography = Typography(),
-        content = content
-    )
-}
