--- app/src/main/java/com/whitebeach/atleticolineupapp/presentations/main/FireStoreViewModel.kt
@@ -1,24 +1,0 @@
-package com.whitebeach.atleticolineupapp.presentations.main
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.whitebeach.data.model.PlayerInfo
-import com.whitebeach.data.remote.firestore.playerFireStoreDataSource
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-
-class FireStoreViewModel() : ViewModel() {
-
-    var playerDataList: StateFlow<List<PlayerInfo>> = MutableStateFlow(emptyList())
-
-    init {
-        getData()
-    }
-
-    fun getData() {
-        viewModelScope.launch {
-            playerDataList = playerFireStoreDataSource()
-        }
-    }
-}
