+++ presentation/src/main/java/com/whitebeach/presentation/component/dragDrop/DragData.kt
@@ -1,0 +1,8 @@
+package com.whitebeach.presentation.component.dragDrop
+
+import com.whitebeach.atleticolineupapp.dragAndDrop.MimeType
+
+class DragData (
+    val type: MimeType = MimeType.TEXT_PLAIN,
+    val data: Any? = null
+)
