+++ presentation/src/main/java/com/whitebeach/presentation/theme/Colorx.kt
@@ -1,0 +1,68 @@
+package com.whitebeach.presentation.theme
+
+import androidx.compose.ui.graphics.Color
+
+val md_theme_light_primary = Color(0xFFC0001F)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFFFDAD7)
+val md_theme_light_onPrimaryContainer = Color(0xFF410004)
+val md_theme_light_secondary = Color(0xFF006874)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFF97F0FF)
+val md_theme_light_onSecondaryContainer = Color(0xFF001F24)
+val md_theme_light_tertiary = Color(0xFF005BBE)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFD7E2FF)
+val md_theme_light_onTertiaryContainer = Color(0xFF001A40)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF410005)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF410005)
+val md_theme_light_surfaceVariant = Color(0xFFF5DDDB)
+val md_theme_light_onSurfaceVariant = Color(0xFF534342)
+val md_theme_light_outline = Color(0xFF857371)
+val md_theme_light_inverseOnSurface = Color(0xFFFFEDEB)
+val md_theme_light_inverseSurface = Color(0xFF5F1416)
+val md_theme_light_inversePrimary = Color(0xFFFFB3AE)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFFC0001F)
+val md_theme_light_outlineVariant = Color(0xFFD8C2C0)
+val md_theme_light_scrim = Color(0xFF000000)
+
+val md_theme_dark_primary = Color(0xFFFFB3AE)
+val md_theme_dark_onPrimary = Color(0xFF68000C)
+val md_theme_dark_primaryContainer = Color(0xFF930015)
+val md_theme_dark_onPrimaryContainer = Color(0xFFFFDAD7)
+val md_theme_dark_secondary = Color(0xFF4FD8EB)
+val md_theme_dark_onSecondary = Color(0xFF00363D)
+val md_theme_dark_secondaryContainer = Color(0xFF004F58)
+val md_theme_dark_onSecondaryContainer = Color(0xFF97F0FF)
+val md_theme_dark_tertiary = Color(0xFFACC7FF)
+val md_theme_dark_onTertiary = Color(0xFF002F67)
+val md_theme_dark_tertiaryContainer = Color(0xFF004491)
+val md_theme_dark_onTertiaryContainer = Color(0xFFD7E2FF)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF410005)
+val md_theme_dark_onBackground = Color(0xFFFFDAD7)
+val md_theme_dark_surface = Color(0xFF410005)
+val md_theme_dark_onSurface = Color(0xFFFFDAD7)
+val md_theme_dark_surfaceVariant = Color(0xFF534342)
+val md_theme_dark_onSurfaceVariant = Color(0xFFD8C2C0)
+val md_theme_dark_outline = Color(0xFFA08C8B)
+val md_theme_dark_inverseOnSurface = Color(0xFF410005)
+val md_theme_dark_inverseSurface = Color(0xFFFFDAD7)
+val md_theme_dark_inversePrimary = Color(0xFFC0001F)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFFFB3AE)
+val md_theme_dark_outlineVariant = Color(0xFF534342)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+
