+++ presentation/src/main/res/drawable/atletico_logo.png
