+++ presentation/src/main/res/drawable/pitch.jpg
