+++ presentation/src/main/res/values/strings.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">AtleticoLineup</string>
+    <string name="ad_id_banner">ca-app-pub-3940256099942544/6300978111</string>
+    //ca-app-pub-3512477696171901/2518004744
+    <string name="loading_failed">Failed to load</string>
+    <string name="retry">Retry</string>
+    <string name="loading">Loading</string>
+</resources>
